ADD CONTAMINANT FILTERING MODULES
This commit adds three modules to filter conatminants from the smRNA-Seq data.
Using databses provided by the user the reads are filtered for rrna, trna, cdna,
ncrna, and pirna. The provided files are filtered to remove miRNAs, indexed
and used as reference to map the reads to using bowtie2. Reads that map to a
contaminant database are removed from the sample.

The process is optional and only those contaminants will be filtered for which
a reference database has been provided.